Created use state const for user favorite movies
